finish up typescriptization and trimming of helpBubble code
