+++ jsapp/js/components/support/helpBubble.scss
@@ -79,7 +79,6 @@ $s-help-bubble-row-spacing: 20px;
   }
 
   .k-icon {
-    font-size: 24px;
     margin: 3px;
     display: block;
   }
